customRuntime/openwhisk-runtime-python/core/python2Action/libRc/requirements.txt
requests
python-swiftclient==3.0.0